SHT11 to DHT11, Fixed message queue's
The better SHT11 sensor REALLY doesn't like long wires, so we have had to move to the DHT11. I've also fixed the message sending/processing. We obay retry limits, and message queue processing better. Messaged are queued by QueueAlerts() and sent by SendAlerts. the string is made by BuildAlertString - This is so we can use it to alert on other outputs too (Serial) - to come...